Initial setup: add packages; create project structure; remove unused files
