+++ src/App.jsx
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <div className="App">
+      App component
+    </div>
+  );
+}
+
+export default App;
